add GUI
I have started working on a simple GUI for the touch screen of the robot.
So far it is not doing anything.
The gui is based on a grid consisting of rows and columns. Each square created by theese parameters
can contain a button. Each grid of buttons will be handled on a higher abstractionlevel not yet defined.
So far basic functions for detecting when a button is klicked and highlighting a button when
holding on it. Also drawing and positioning functions has been implemented.

Due note, none of theese functions are done yet.


Former-commit-id: 24f979dd98f631cdf0c4e669cd538854039fcc43